fix(main): assign unique lesson kind badge tones
